+++ bench.rb
@@ -5,10 +5,20 @@ require "fsck/deep"
 
 N = 1_000
 
-Benchmark.bm(10) do |r|
+Benchmark.bmbm(10) do |r|
   r.report("vanilla") { N.times { 42.succ } }
   r.report("fscked") { N.times { 42.gimme_succ_please } }
 end
+
+# v0.0.4
+# Rehearsal ---------------------------------------------
+# vanilla     0.000000   0.000000   0.000000 (  0.000163)
+# fscked      0.000000   0.000000   0.000000 (  0.003473)
+# ------------------------------------ total: 0.000000sec
+# 
+#                 user     system      total        real
+# vanilla     0.000000   0.000000   0.000000 (  0.000100)
+# fscked      0.000000   0.000000   0.000000 (  0.000199)
 
 # v0.0.3
 #                 user     system      total        real